+++ myflask_app/config.py
@@ -1,4 +1,4 @@
-#Centralized configuration settings.
-# config.py
-SECRET_KEY = 'your_secret_key'
-SQLALCHEMY_DATABASE_URI = 'mysql://root:password123@localhost/temporary_user_db'
+# Configuration settings (if needed)
+SECRET_KEY = "H3lpH!v3"
+SQLALCHEMY_DATABASE_URI = "mysql+mysqldb://root:password123@localhost/temporary_user_db"
+SQLALCHEMY_TRACK_MODIFICATIONS = False
