+++ myflask_app/exceptions/custom_exceptions.py
@@ -1,1 +1,3 @@
 #Custom exceptions for the app.
+class CustomException(Exception):
+    pass
