+++ myflask_app/middleware/request_middleware.py
@@ -1,1 +1,4 @@
 #Functions that handle before and after request actions.
+def custom_middleware(request):
+    # Custom request middleware logic
+    pass
