+++ myflask_app/services/user_services.py
@@ -1,0 +1,3 @@
+def some_custom_logic(data):
+    # Custom logic for module 1
+    pass
