+++ myflask_app/tests/test_module1_resources.py
@@ -1,1 +1,7 @@
 #Unit tests for your resources.
+import unittest
+from app import app
+
+class Module1Tests(unittest.TestCase):
+    # Unit tests for module 1
+    pass
